requirements.txt
alembic==1.7.7
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.3
apturl==0.5.2
argcomplete==2.0.0
asn1crypto==1.5.1
attrs==21.4.0
bcrypt==3.1.7
bitarray==1.2.2
blake2b-py==0.1.4
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.15.0
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
coincurve==17.0.0
colorama==0.4.3
coloredlogs==15.0.1
command-not-found==0.3
configparser==5.2.0
coverage==6.3.2
cryptography==2.8
crytic-compile==0.2.2
cupshelpers==1.0
cycler==0.11.0
cytoolz==0.11.2
dbus-python==1.2.16
defer==1.0.6
dictionaries==0.0.2
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
docker==5.0.3
docker-pycreds==0.4.0
duplicity==0.8.12.0
email-validator==1.1.3
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.7
eth-bloom==1.0.4
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
ethereum-input-decoder==0.2.2
fasteners==0.14.1
filelock==3.6.0
Flask==1.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.31.2
future==0.18.2
gitdb==4.0.9
gitdb2==2.0.6
GitPython==3.1.27
greenlet==1.1.2
gunicorn==20.0.4
hexbytes==0.2.2
httplib2==0.14.0
humanfriendly==10.0
identify==2.4.12
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==4.4.0
keyring==18.0.1
kiwisolver==1.4.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lru-dict==1.1.7
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
matplotlib==3.5.1
mock==4.0.3
monotonic==1.5
more-itertools==4.2.0
mypy-extensions==0.4.3
mythril==0.22.42
netifaces==0.10.4
nodeenv==1.6.0
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
parsimonious==0.8.1
pbr==5.8.1
persistent==4.9.0
pexpect==4.6.0
Pillow==7.0.0
pipenv==11.9.0
pipx==1.0.0
platformdirs==2.5.1
pluggy==1.0.0
pre-commit==2.17.0
prettytable==3.2.0
protobuf==3.6.1
py==1.11.0
py-ecc==4.1.0
py-evm==0.5.0a1
py-flags==1.1.4
py-solc==3.2.0
py-solc-x==1.1.1
pycairo==1.16.2
pycparser==2.21
pycryptodome==3.14.1
pycups==1.9.73
pyethash==0.1.27
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pysha3==1.0.2
pytest==7.1.1
pytest-cov==3.0.0
pytest-mock==3.7.0
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.5
pytz==2019.3
pyxdg==0.26
PyYAML==6.0
reportlab==3.5.34
requests==2.27.1
requests-unixsocket==0.2.0
rlp==2.0.1
sarif-om==1.0.4
SecretStorage==2.3.1
semantic-version==2.9.0
simplejson==3.16.0
six==1.14.0
slither-analyzer==0.8.2
smmap==5.0.0
smmap2==2.0.5
solidity-parser==0.1.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.29
systemd-python==234
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
transaction==3.0.1
trie==2.0.0a5
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
usb-creator==0.3.7
userpath==1.8.0
virtualenv==20.14.0
virtualenv-clone==0.3.0
wadllib==1.3.3
wcwidth==0.2.5
websocket-client==1.3.1
Werkzeug==1.0.1
wincertstore==0.2
WTForms==3.0.1
xkit==0.0.0
z3-solver==4.8.15.0
zipp==3.8.0
zope.interface==5.4.0
